complete add and show profile picture function

+++ flaskblog/views/user_management.py
@@ -1,9 +1,15 @@
-from flask import Blueprint
-from flask import render_template, flash, request
+from flask import Blueprint, current_app
+from flask import render_template, flash, request, redirect, url_for
 from werkzeug.security import generate_password_hash
-from flask_login import login_required
+from flask_login import login_required, current_user
 from webforms import UserForm, NamerForm
 from models import db, Users
+# to ensure secure image filename
+from werkzeug.utils import secure_filename
+import uuid as uuid
+import os
+# read and write image using path
+import cv2
 
 manage_users = Blueprint('user_management', __name__)
 
@@ -78,17 +84,45 @@ def update_user(id):
 
     user_to_update = Users.query.get_or_404(id)
     if request.method == 'POST':
+        user_previous_pic = user_to_update.profile_pic
         user_to_update.username = request.form['username']
         user_to_update.name = request.form['name']
         user_to_update.email = request.form['email']
         user_to_update.mobile = request.form['mobile']
         user_to_update.about_author = request.form['about_author']
-        try:
-            # save into database
-            db.session.commit()
-            flash("User Updated Successfully")
-        except:
-            flash("User Not Found")
+        if request.files['profile_pic']:
+            profile_picture = request.files['profile_pic']
+            pic_filename = secure_filename(profile_picture.filename)
+            # set UUID to make image filename unique
+            pic_name = str(uuid.uuid1()) + "_" + pic_filename
+        # else: 
+        #     # use default image as user profile picture
+        #     profile_picture = cv2.imread(os.path.join(current_app.config['UPLOAD_FOLDER'], 'user.png'))
+        #     # set UUID to make image filename unique
+        #     pic_name = str(uuid.uuid1()) + "_" + "user.png"
+        # grab image filename
+            user_to_update.profile_pic = pic_name
+            try:
+                # save into database
+                db.session.commit()
+                # remove previous image
+                if os.path.exists(os.path.join(current_app.config['UPLOAD_FOLDER'],user_previous_pic)):
+                    os.remove(os.path.join(current_app.config['UPLOAD_FOLDER'], user_previous_pic))
+                # save the image
+                # if request.files['profile_pic']:
+                profile_picture.save(os.path.join(current_app.config['UPLOAD_FOLDER'], pic_name))
+                # else:
+                #     cv2.imwrite(os.path.join(current_app.config['UPLOAD_FOLDER'], pic_name), profile_picture)
+                flash("User Updated Successfully")
+            except:
+                flash("User Not Found")
+        
+        else:
+            try:
+                db.session.commit()
+                flash("User Updated Successfully")
+            except:
+                flash("User Not Found")
 
     context = {
         'form': form,
@@ -99,21 +133,25 @@ def update_user(id):
 @manage_users.route('/delete/<int:id>')
 @login_required
 def delete_user(id):
-    form = UserForm()
+    if id == current_user.id:
+        form = UserForm()
 
-    user_to_delete = Users.query.get_or_404(id)
+        user_to_delete = Users.query.get_or_404(id)
 
-    try:
-        db.session.delete(user_to_delete)
-        db.session.commit()
-        flash("User Deleted Successfully")
-    except:
-        flash("User Not Found")
+        try:
+            db.session.delete(user_to_delete)
+            db.session.commit()
+            flash("User Deleted Successfully")
+        except:
+            flash("User Not Found")
 
-    user_list = Users.query.order_by(Users.date_added)
+        user_list = Users.query.order_by(Users.date_added)
 
-    context = {
-        'form': form,
-        'user_list': user_list
-    }
-    return render_template('user_management/add_user.html', **context)
+        context = {
+            'form': form,
+            'user_list': user_list
+        }
+        return render_template('user_management/add_user.html', **context)
+    else:
+        flash("You can't delete the user")
+        return redirect(url_for('dashboard'))
